pyproject.toml
[tool.poetry]
name = "fccu"
version = "1.0.0"
description = "ОАО Финансовая Компания Кредитных Союзов"
authors = ["Jorupbek <a.jorupbek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
django = "^4.2.5"
python = "^3.11"
flake8 = "^6.1.0"
pylint-django = "^2.5.3"
requests = "^2.31.0"
pillow = "^10.0.0"
django-jazzmin = "^2.6.0"
django-widget-tweaks = "^1.5.0"
psycopg2-binary = "^2.9.7"
gunicorn = "^21.2.0"
django-environ = "^0.11.2"
bs4 = "^0.0.1"
django-ckeditor = "^6.7.0"
django-admin-soft-dashboard = "^1.0.12"
reportlab = "^4.0.5"
python-docx = "^0.8.11"
django-tinymce = "^3.6.1"
docxtpl = "^0.16.7"
django-crispy-forms = "^2.1"
crispy-bootstrap4 = "^2022.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
